Fix int64 overflows

+++ wavemodel.go
@@ -129,9 +129,9 @@ func GetWaveModelForLocation(loc Location) *WaveModel {
 }
 
 // Get the time and hour of the latest NOAA WaveWatch model run
-func LatestModelDateTime() (time.Time, int) {
+func LatestModelDateTime() (time.Time, int64) {
 	currentTime := time.Now().Local()
 	lastModelHour := currentTime.Hour() - (currentTime.Hour() % 6)
-	currentTime = currentTime.Add(time.Duration(-(currentTime.Hour() % 6) * int(time.Hour)))
+	currentTime = currentTime.Add(time.Duration(-(currentTime.Hour() % 6) * int64(time.Hour)))
 	return currentTime, lastModelHour
 }
